resolved feedback and refactor prism component import
