--- components/prism/docker.ts
@@ -1,13 +1,0 @@
-export default function PrismDocker(Prism: any) {
-  Prism.languages.docker = {
-    keyword: {
-      pattern: /(^\s*)(?:ADD|ARG|CMD|COPY|ENTRYPOINT|ENV|EXPOSE|FROM|HEALTHCHECK|LABEL|MAINTAINER|ONBUILD|RUN|SHELL|STOPSIGNAL|USER|VOLUME|WORKDIR)(?=\s)/im,
-      lookbehind: true,
-    },
-    string: /("|')(?:(?!\1)[^\\\r\n]|\\(?:\r\n|[\s\S]))*\1/,
-    comment: /#.*/,
-    punctuation: /---|\.\.\.|[:[\]{}\-,|>?]/,
-  };
-
-  Prism.languages.dockerfile = Prism.languages.docker;
-}
